feat: Move footer from custom-app to fern/components
