--- custom-app/src/assets/AlchemyUniversityIcon.tsx
@@ -1,42 +1,0 @@
-import type { FC } from 'react'
-
-interface AlchemyUniversityIconProps {
-  isDark?: boolean
-}
-
-const AlchemyUniversityIcon: FC<AlchemyUniversityIconProps> = ({
-  isDark = false,
-}) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-    >
-      <path
-        d="M14.8223 13.4397V2.00037L12.3848 2V3.07335H13.4866V12.3663H12.3848V13.4397H14.8223Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-      <path
-        d="M2.00227 12.3667V3.07333L3.09827 3.07369V2.00035H0.666626V13.44H3.09827V12.3667H2.00227Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-      <path
-        d="M7.79621 4.21892L10.4846 8.74039C10.4979 8.76271 10.5047 8.78796 10.5047 8.81344C10.5045 8.83892 10.4975 8.86417 10.4844 8.88626C10.4711 8.90836 10.4522 8.92685 10.4294 8.9397C10.4066 8.95255 10.3807 8.95931 10.3543 8.95931H8.74406C8.6913 8.95931 8.63944 8.94579 8.59367 8.92031C8.54813 8.89461 8.51003 8.85786 8.48365 8.81344L6.73067 5.86391C6.70429 5.81949 6.69031 5.76922 6.69031 5.71804C6.69031 5.66685 6.70429 5.61658 6.73067 5.57216L7.5358 4.21734C7.5491 4.19525 7.56804 4.17676 7.59104 4.16413C7.61403 4.15151 7.63996 4.14497 7.66634 4.14497C7.69272 4.14497 7.71865 4.15196 7.74165 4.16504C7.76442 4.17789 7.78313 4.1966 7.79621 4.21892Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-      <path
-        d="M6.1246 6.58215L3.43617 11.1036C3.42264 11.1259 3.41565 11.1512 3.41565 11.1769C3.41565 11.2028 3.42241 11.2281 3.43572 11.2504C3.44879 11.2727 3.46796 11.2912 3.49095 11.3041C3.51395 11.3169 3.53988 11.3235 3.56648 11.3235H5.17742C5.23018 11.3235 5.28181 11.3101 5.32758 11.2844C5.37313 11.2587 5.41101 11.222 5.43738 11.1776L7.19037 8.22804C7.21675 8.18363 7.2305 8.13335 7.2305 8.08194C7.2305 8.03054 7.21675 7.98026 7.19037 7.93584L6.38478 6.58215C6.37148 6.56006 6.35254 6.54179 6.32977 6.52894C6.307 6.51609 6.28107 6.50933 6.25469 6.50933C6.22831 6.50933 6.20238 6.51609 6.17961 6.52894C6.15684 6.54157 6.1379 6.56006 6.1246 6.58215Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-      <path
-        d="M6.386 11.3223H11.7629C11.7892 11.3223 11.8152 11.3155 11.8379 11.3027C11.8607 11.2898 11.8797 11.2713 11.893 11.2492C11.906 11.2269 11.913 11.2019 11.913 11.1762C11.913 11.1505 11.906 11.1255 11.8927 11.1031L11.0885 9.74877C11.0621 9.70435 11.0242 9.6676 10.9785 9.6419C10.9327 9.6162 10.8808 9.6029 10.8281 9.6029H7.32167C7.26892 9.6029 7.21706 9.6162 7.17129 9.6419C7.12552 9.6676 7.08764 9.70435 7.06126 9.74877L6.25613 11.1031C6.24283 11.1252 6.23584 11.1505 6.23584 11.1762C6.23584 11.2019 6.2426 11.2269 6.25591 11.2492C6.26898 11.2713 6.28792 11.2898 6.31092 11.3027C6.33369 11.3155 6.35962 11.3223 6.386 11.3223Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-    </svg>
-  )
-}
-
-export default AlchemyUniversityIcon
