--- custom-app/src/assets/DiscordIcon.tsx
@@ -1,24 +1,0 @@
-import type { FC } from 'react'
-
-interface DiscordIconProps {
-  isDark?: boolean
-}
-
-const DiscordIcon: FC<DiscordIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-    >
-      <path
-        d="M12.6556 3.56509C11.7258 3.13937 10.7441 2.83735 9.73575 2.66675C9.59776 2.91342 9.47291 3.16721 9.36172 3.42705C8.28762 3.2652 7.19533 3.2652 6.12123 3.42705C6.00998 3.16723 5.88513 2.91345 5.7472 2.66675C4.73819 2.83879 3.75589 3.14153 2.82508 3.56732C0.977194 6.3013 0.47626 8.96738 0.726727 11.5956C1.8089 12.3952 3.02016 13.0033 4.30785 13.3934C4.5978 13.0034 4.85437 12.5897 5.07484 12.1567C4.65609 12.0003 4.25193 11.8073 3.86703 11.58C3.96833 11.5066 4.0674 11.4309 4.16314 11.3574C5.28311 11.8841 6.50551 12.1572 7.74314 12.1572C8.98079 12.1572 10.2032 11.8841 11.3232 11.3574C11.42 11.4364 11.5191 11.5121 11.6193 11.58C11.2336 11.8077 10.8287 12.001 10.4092 12.1578C10.6294 12.5906 10.886 13.004 11.1762 13.3934C12.465 13.0048 13.6772 12.397 14.7596 11.5967C15.0534 8.54883 14.2575 5.90723 12.6556 3.56509ZM5.39209 9.97927C4.69413 9.97927 4.1175 9.34587 4.1175 8.56664C4.1175 7.78741 4.67409 7.14844 5.38987 7.14844C6.10565 7.14844 6.67782 7.78741 6.66558 8.56664C6.65334 9.34587 6.10342 9.97927 5.39209 9.97927ZM10.0942 9.97927C9.39511 9.97927 8.82071 9.34587 8.82071 8.56664C8.82071 7.78741 9.3773 7.14844 10.0942 7.14844C10.8111 7.14844 11.3788 7.78741 11.3666 8.56664C11.3543 9.34587 10.8055 9.97927 10.0942 9.97927Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-    </svg>
-  )
-}
-
-export default DiscordIcon
