--- custom-app/src/assets/EmailIcon.tsx
@@ -1,26 +1,0 @@
-import type { FC } from 'react'
-
-interface EmailIconProps {
-  isDark?: boolean
-}
-
-const EmailIcon: FC<EmailIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-    >
-      <path
-        d="M2.79998 4.39995L7.54461 7.6847C7.81856 7.87436 8.18139 7.87436 8.45534 7.6847L13.2 4.39995M3.19998 12.8H12.8C13.6836 12.8 14.4 12.0836 14.4 11.2V4.79995C14.4 3.9163 13.6836 3.19995 12.8 3.19995H3.19998C2.31632 3.19995 1.59998 3.9163 1.59998 4.79995V11.2C1.59998 12.0836 2.31632 12.8 3.19998 12.8Z"
-        stroke={isDark ? '#FBFDFF' : '#475569'}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-export default EmailIcon
