--- custom-app/src/assets/NewsletterIcon.tsx
@@ -1,27 +1,0 @@
-import type { FC } from 'react'
-
-interface NewsletterIconProps {
-  isDark?: boolean
-}
-
-const NewsletterIcon: FC<NewsletterIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      fill="none"
-    >
-      <path
-        d="M4.875 3.75012C3 3.75012 1.5 5.25012 1.5 7.12512V12.7501C1.5 13.1479 1.65804 13.5295 1.93934 13.8108C2.22064 14.0921 2.60218 14.2501 3 14.2501H15C15.3978 14.2501 15.7794 14.0921 16.0607 13.8108C16.342 13.5295 16.5 13.1479 16.5 12.7501V6.75012C16.5 5.10012 15.15 3.75012 13.5 3.75012H4.875ZM4.875 3.75012C6.75 3.75012 8.25 5.25012 8.25 7.12512V12.7501C8.25 13.1479 8.09196 13.5295 7.81066 13.8108C7.52936 14.0921 7.14782 14.2501 6.75 14.2501M11.25 6.75012H13.5V8.25012M4.5 7.50012H5.25"
-        stroke={isDark ? '#FBFDFF' : '#475569'}
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-export default NewsletterIcon
