--- custom-app/src/assets/RobotIcon.tsx
@@ -1,27 +1,0 @@
-import type { FC } from 'react'
-
-interface RobotIconProps {
-  isDark?: boolean
-}
-
-const RobotIcon: FC<RobotIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      fill="none"
-    >
-      <path
-        d="M9 6.00012V3.00012H6M1.5 10.5001H3M15 10.5001H16.5M11.25 9.75012V11.2501M6.75 9.75012V11.2501M4.5 6.00012H13.5C14.3284 6.00012 15 6.67169 15 7.50012V13.5001C15 14.3285 14.3284 15.0001 13.5 15.0001H4.5C3.67157 15.0001 3 14.3285 3 13.5001V7.50012C3 6.67169 3.67157 6.00012 4.5 6.00012Z"
-        stroke={isDark ? '#FBFDFF' : '#475569'}
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-export default RobotIcon
