--- custom-app/src/assets/SupportHubIcon.tsx
@@ -1,26 +1,0 @@
-import type { FC } from 'react'
-
-interface SupportHubIconProps {
-  isDark?: boolean
-}
-
-const SupportHubIcon: FC<SupportHubIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-    >
-      <path
-        d="M9.69706 9.69712C8.7598 10.6344 7.2402 10.6344 6.30294 9.69712M9.69706 9.69712C10.6343 8.75986 10.6343 7.24026 9.69706 6.30301M9.69706 9.69712L12.2426 12.2427M6.30294 9.69712C5.36569 8.75986 5.36569 7.24026 6.30294 6.30301M6.30294 9.69712L3.75736 12.2427M6.30294 6.30301C7.2402 5.36575 8.7598 5.36575 9.69706 6.30301M6.30294 6.30301L3.75736 3.75742M9.69706 6.30301L12.2426 3.75742M12.5255 12.5255C10.0261 15.0249 5.97387 15.0249 3.47452 12.5255C0.975162 10.0262 0.975162 5.97394 3.47452 3.47458C5.97387 0.975224 10.0261 0.975224 12.5255 3.47458C15.0248 5.97394 15.0248 10.0262 12.5255 12.5255Z"
-        stroke={isDark ? '#FBFDFF' : '#475569'}
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-export default SupportHubIcon
