--- custom-app/src/assets/XIcon.tsx
@@ -1,24 +1,0 @@
-import type { FC } from 'react'
-
-interface XIconProps {
-  isDark?: boolean
-}
-
-const XIcon: FC<XIconProps> = ({ isDark = false }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-    >
-      <path
-        d="M9.15275 6.92804L14.2525 1H13.044L8.61591 6.14724L5.07919 1H1L6.34821 8.78354L1 15H2.20855L6.88474 9.56434L10.6198 15H14.699L9.15245 6.92804H9.15275ZM7.49748 8.85211L6.95559 8.07705L2.644 1.90978H4.50026L7.97976 6.88696L8.52165 7.66202L13.0446 14.1316H11.1883L7.49748 8.85241V8.85211Z"
-        fill={isDark ? '#FBFDFF' : '#475569'}
-      />
-    </svg>
-  )
-}
-
-export default XIcon
