+++ custom-app/src/main.tsx
@@ -1,64 +1,41 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 
-import { CustomFooter } from './Footer.js'
 import { Codeblock } from './Codeblock.js'
 
 const FERN_FOOTER_CONTAINER_ID = 'fern-footer'
 
 const render = async () => {
-  const alchemyFooterId = document.getElementById('alchemy-footer')
-
-  if (!alchemyFooterId) {
-    // Create alchemy footer wrapper with a data attribute to help with hydration
-    const alchemyContentWrapper = document.createElement('div')
-    alchemyContentWrapper.setAttribute('id', 'alchemy-footer')
-    alchemyContentWrapper.setAttribute('data-react-component', 'true')
-
-    // Get or create fern-footer container
-    let fernFooterContainer = document.getElementById(FERN_FOOTER_CONTAINER_ID)
-    if (!fernFooterContainer) {
-      fernFooterContainer = document.createElement('div')
-      fernFooterContainer.setAttribute('id', FERN_FOOTER_CONTAINER_ID)
-      document.body.appendChild(fernFooterContainer)
-    }
-
-    // Get or create code-block-id container
-    let codeBlockContainer = document.getElementById('code-block-id')
-    if (!codeBlockContainer) {
-      codeBlockContainer = document.createElement('div')
-      codeBlockContainer.setAttribute('id', 'code-block-id')
-      document.body.appendChild(codeBlockContainer)
-    }
-
-    fernFooterContainer.insertBefore(
-      alchemyContentWrapper,
-      fernFooterContainer.firstChild,
-    )
-
-    // Use createRoot instead of render for React 18+ compatibility
-    const root = createRoot(alchemyContentWrapper)
-    root.render(
-      <React.StrictMode>
-        <CustomFooter />
-      </React.StrictMode>,
-    )
-
-    // Render Codeblock into code-block-id container
-    const codeBlockRoot = createRoot(codeBlockContainer)
-    codeBlockRoot.render(
-      <React.StrictMode>
-        <Codeblock />
-      </React.StrictMode>,
-    )
-
-    // Show the container after rendering
-    if (fernFooterContainer) fernFooterContainer.style.display = 'block'
-    if (codeBlockContainer) codeBlockContainer.style.display = 'block'
-
-    // Remove the default built with fern link since we add one in the Footer component.
-    document.getElementById('builtwithfern')?.remove()
+  // Get or create fern-footer container
+  let fernFooterContainer = document.getElementById(FERN_FOOTER_CONTAINER_ID)
+  if (!fernFooterContainer) {
+    fernFooterContainer = document.createElement('div')
+    fernFooterContainer.setAttribute('id', FERN_FOOTER_CONTAINER_ID)
+    document.body.appendChild(fernFooterContainer)
   }
+
+  // Get or create code-block-id container
+  let codeBlockContainer = document.getElementById('code-block-id')
+  if (!codeBlockContainer) {
+    codeBlockContainer = document.createElement('div')
+    codeBlockContainer.setAttribute('id', 'code-block-id')
+    document.body.appendChild(codeBlockContainer)
+  }
+
+  // Render Codeblock into code-block-id container
+  const codeBlockRoot = createRoot(codeBlockContainer)
+  codeBlockRoot.render(
+    <React.StrictMode>
+      <Codeblock />
+    </React.StrictMode>,
+  )
+
+  // Show the container after rendering
+  if (fernFooterContainer) fernFooterContainer.style.display = 'block'
+  if (codeBlockContainer) codeBlockContainer.style.display = 'block'
+
+  // Remove the default built with fern link since we add one in the Footer component.
+  document.getElementById('builtwithfern')?.remove()
 }
 
 // Use 'load' event instead of 'DOMContentLoaded' for App Router
@@ -82,11 +59,6 @@ window.addEventListener('load', async () => {
       )
     if (shouldRender) {
       await render()
-    }
-
-    // Optionally: Remove custom footer if not on /docs
-    if (currentPath !== '/docs' && document.getElementById('alchemy-footer')) {
-      document.getElementById('alchemy-footer')?.remove()
     }
   }).observe(document.body, { childList: true, subtree: true })
 })
