+++ fern/assets/styles.css
@@ -762,3 +762,4 @@ header#fern-header a.fern-button.filled.normal.primary:hover {
     }
   }
 }
+
