+++ fern/components/Footer.css
@@ -1,0 +1,178 @@
+html.light #AlchemyFooter {
+  --border-color: #d4d4e8;
+  --color: #020617;
+}
+
+html.dark #AlchemyFooter {
+  --border-color: #3f3f40;
+  --color: #fbfdff;
+}
+
+#AlchemyFooter {
+  width: 100%;
+
+  .FooterContainer {
+    padding: 48px 24px 54px;
+    color: var(--color);
+  }
+
+  .FooterWrapper {
+    max-width: 1010px;
+    width: 100%;
+    margin: 0 auto;
+    display: flex;
+    flex-direction: column;
+    padding: 0 24px;
+  }
+
+  .TopSection {
+    padding: 48px 0px;
+    border-top: 1px solid;
+    border-bottom: 1px solid;
+    border-color: var(--border-color);
+    display: flex;
+  }
+
+  .LinksColumns {
+    display: flex;
+    flex-direction: column;
+    row-gap: 36px;
+
+    @media screen and (min-width: 768px) {
+      flex-direction: row;
+      column-gap: 64px;
+      justify-content: space-between;
+    }
+  }
+
+  .Column {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    font-size: 14px;
+  }
+
+  .ColumnTitle {
+    font-size: 16px;
+    font-weight: 600;
+    line-height: 1.25;
+    letter-spacing: -0.32px;
+    margin: 0;
+  }
+
+  .FooterLink {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    text-decoration: none;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 1.5;
+    &:hover {
+      color: #5167ff;
+
+      svg {
+        /* Add hover effect to icon paths depending on if they use fill or stroke */
+        path[fill] {
+          fill: #5167ff;
+        }
+        path[stroke] {
+          stroke: #5167ff;
+        }
+      }
+    }
+  }
+
+  .BottomSection {
+    justify-content: center;
+    padding-top: 18px;
+    font-size: 14px;
+    font-weight: 500;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    gap: 28px;
+    margin-top: 36px;
+
+    @media screen and (min-width: 768px) {
+      flex-direction: row;
+      margin-top: 0;
+    }
+  }
+
+  .CopyrightTerms {
+    display: flex;
+    align-items: center;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 1.5;
+    flex-direction: column;
+    gap: 12px;
+
+    @media screen and (min-width: 768px) {
+      flex-direction: row;
+      gap: 24px;
+      width: 100%;
+      justify-content: space-between;
+    }
+  }
+
+  .LeftTerms {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 12px;
+
+    @media screen and (min-width: 768px) {
+      flex-direction: row;
+      align-items: center;
+      gap: 24px;
+    }
+  }
+
+  .RightTerms {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-top: 12px;
+
+    @media screen and (min-width: 768px) {
+      margin-top: 0;
+      justify-content: flex-end;
+    }
+  }
+
+  .TermsLink {
+    text-decoration: none;
+    &:hover {
+      color: #5167ff;
+    }
+  }
+
+  .LogoColumn {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    font-size: 14px;
+
+    @media screen and (max-width: 768px) {
+      margin-bottom: 12px;
+    }
+  }
+
+  .FernLogo {
+    .transition {
+      transition:
+        fill 0.2s,
+        stroke 0.2s;
+    }
+    &:hover svg path:last-of-type {
+      fill: #1ba32a !important;
+      stroke: #1ba32a !important;
+    }
+    &:hover svg path:first-of-type {
+      fill: currentColor;
+      stroke: currentColor;
+    }
+  }
+}
