+++ fern/docs.yml
@@ -71,6 +71,7 @@ typography:
 css:
   - ./assets/styles.css
   - ./assets/styles-home.css
+  - ./components/Footer.css
 
 js:
   - path: ./dist/output.js
@@ -82,6 +83,8 @@ js:
 
 experimental:
   openapi-parser-v3: true
+  mdx-components:
+    - ./components
 
 # TODO: is it possible to add a Font Awesome icon to a navbar-link?
 navbar-links:
