+++ fern/fern.config.json
@@ -1,4 +1,4 @@
 {
   "organization": "alchemy",
-  "version": "0.64.10"
+  "version": "0.65.34"
 }
