+++ fern/home/index.mdx
@@ -4,6 +4,8 @@ layout: custom
 hide-feedback: true
 ---
 
+import { AlchemyFooter } from "../components/Footer.tsx";
+
 <div className="homepage-content">
   <div className="products">
     <h2 id="build-anything-onchain" data-state="closed">Build anything onchain</h2>
@@ -127,6 +129,8 @@ hide-feedback: true
       </div>
     </div>
   </div>
+
+  <AlchemyFooter />
 
   {/* Hide this section for now as the new Footer will cover it and this will be repurposed. */}
   {/* <div className="footer">
