+++ package.json
@@ -38,7 +38,7 @@
   "dependencies": {
     "@apidevtools/json-schema-ref-parser": "^12.0.2",
     "@open-rpc/schema-utils-js": "^2.1.2",
-    "fern-api": "^0.64.10",
+    "fern-api": "^0.65.34",
     "json-schema-merge-allof": "^0.8.1",
     "ts-node": "^10.9.2"
   },
