+++ tsconfig.json
@@ -4,6 +4,7 @@
     "module": "NodeNext",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
+    "jsx": "react",
     "strict": true,
     "skipLibCheck": true
   },
